consolidate "common"
